re-write webpack config
